Test the other overloads.
x and x= currently return a plain value: it should be easy to make them
defer, though maybe not very cheap. It may be worth providing a complete
set of string ops at some point, with more structure inside the object
to support them, but that's a lot more work.

++ and -- appear to numify the object to its refaddr before operating.
This is so odd I think it must be a bug in perl: it also happens with an
object with a trivial stringify overload. If this doesn't get fixed (and
for perls without the fix, if it does) I think it would be better to
provide explicit overloads for these two which croak.

.= needs tests in defer.t, and possibly also an implementation.